fix OverflowException

+++ src/ui/Helpers/FieldHelper.cs
@@ -13,7 +13,7 @@ namespace GovUk.Education.ManageCourses.Ui.Helpers
 
         public static string DisplayText(this decimal? value, int maxLength = 100, string defaultEmpty = "This field is empty")
         {
-            var text = value.HasValue ? string.Format(CultureInfo.InvariantCulture, "£{0:n0}", (int) value.Value) : "";
+            var text = value.HasValue ? string.Format(CultureInfo.InvariantCulture, "£{0:n0}", value.Value) : "";
             var result = DisplayText(text, maxLength, defaultEmpty);
             return result;
         }
